cleanup: remove MustacheCustomTagInterceptor, remove unused mustache custom tags
