+++ src/main/java/alfio/util/TemplateManager.java
@@ -44,7 +44,8 @@ import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static alfio.util.MustacheCustomTagInterceptor.COUNTRY_NAME;
+import static alfio.util.MustacheCustomTag.ADDITIONAL_FIELD_VALUE;
+import static alfio.util.MustacheCustomTag.COUNTRY_NAME;
 
 /**
  * For hiding the ugliness :)
@@ -127,8 +128,9 @@ public class TemplateManager {
     private String render(AbstractResource resource, Map<String, Object> model, Locale locale, TemplateOutput templateOutput) {
         try {
             ModelAndView mv = new ModelAndView((String) null, model);
-            mv.addObject("format-date", MustacheCustomTagInterceptor.FORMAT_DATE);
+            mv.addObject("format-date", MustacheCustomTag.FORMAT_DATE);
             mv.addObject("country-name", COUNTRY_NAME);
+            mv.addObject("additional-field-value", ADDITIONAL_FIELD_VALUE.apply(model.get("additional-fields")));
             mv.addObject(MustacheLocalizationMessageInterceptor.DEFAULT_MODEL_KEY, new CustomLocalizationMessageInterceptor(locale, messageSource).createTranslator());
             return compile(resource, templateOutput).execute(mv.getModel());
         } catch (Exception e) {
